predict: sanitzation

+++ plugins/coreml/src/coreml/__init__.py
@@ -250,13 +250,13 @@ class CoreMLPlugin(PredictPlugin, scrypted_sdk.Settings, scrypted_sdk.DeviceProv
 
             for r in objects:
                 obj = Prediction(
-                    r["classId"].astype(float),
-                    r["confidence"].astype(float),
+                    r["classId"],
+                    r["confidence"],
                     Rectangle(
-                        r["xmin"].astype(float),
-                        r["ymin"].astype(float),
-                        r["xmax"].astype(float),
-                        r["ymax"].astype(float),
+                        r["xmin"],
+                        r["ymin"],
+                        r["xmax"],
+                        r["ymax"],
                     ),
                 )
                 objs.append(obj)
@@ -275,9 +275,9 @@ class CoreMLPlugin(PredictPlugin, scrypted_sdk.Settings, scrypted_sdk.DeviceProv
             ),
         )
 
-        coordinatesList = out_dict["coordinates"].astype(float)
+        coordinatesList = out_dict["coordinates"]
 
-        for index, confidenceList in enumerate(out_dict["confidence"].astype(float)):
+        for index, confidenceList in enumerate(out_dict["confidence"]):
             values = confidenceList
             maxConfidenceIndex = max(range(len(values)), key=values.__getitem__)
             maxConfidence = confidenceList[maxConfidenceIndex]
